+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix an issue where the `ToString` implementation of `Ptr` was causing heap buffer overflow (#11).
 - Fix an issue where `Std::eprint` and `Std::eprintln` wrote the output to stderr, not to stdout (#13).
 - Fixed an issue where exponential notation of integer literals such as 3e10 was not working (#14).
+- Fixed an issue where parsing floating point literal without decimal point (e.g., `1_F32`) causes the compiler to panic.
 
 ## [0.1.0] - 2023-10-24